requirements.txt
torch>=1.9.0
torchvision
torchaudio
numpy
pandas
matplotlib
tqdm
PyYAML
scikit-learn
scipy